Dashboards: Extract PanelDataPane factory to break circular dependency
Move the `createFor` static method from PanelDataPane into a standalone
`createPanelDataPane` factory function in PanelDataPaneFactory.ts. This
breaks the circular dependency between PanelDataPane.tsx and
PanelDataTransformationsTab.tsx — PanelDataPane no longer needs to import
any tab implementations.
